Added Capistrano for easy deploying
Capistrano is a great tool which provides an easy way in which to deploy
to server. I've adapted the instructions from the great tutorial at [1].

Capistrano provides two different environments; staging and production.
We by default want to push to staging, and then we'll manually merge
into production as-and-when we're happy with it.

Closes #24

[1]:
http://www.gashawmola.com/devops/2015/09/07/deploy-a-jekylle-website-using-capistrano-and-nginx.html